Refactor model and run file
